+++ index.php
@@ -201,7 +201,9 @@ $cachesStyleSheets = json_encode(array_values($styles));
 	<?php if ($themeSwitcher) : ?>
 	<script>
 		(() => {
-			const theme = localStorage.getItem('theme') ?? 'is-light'
+			const prefersColourScheme = window.matchMedia('(prefers-color-scheme:light)')
+			const colourScheme = prefersColourScheme.matches ? 'is-light' : 'is-dark'
+			const theme = localStorage.getItem('theme') ?? colourScheme
 			document.documentElement.classList.add(theme)
 
 			const styles = <?php echo $cachesStyleSheets; ?>;